requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
black==19.10b0
Click==7.0
Django==3.0.2
django-markdown-deux==1.0.5
django-pagedown==2.0.3
gunicorn==19.9.0
markdown2==2.3.8
pathspec==0.7.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-decouple==3.1
pytz==2019.1
regex==2020.1.8
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.1
whitenoise==5.0.1
